pyproject.toml
[tool.poetry]
name = "lego-songbook-manager"
version = "2019.12.0"
description = "Tools to easily manage Lego Songbook"
authors = ["Kip Yin <28321392+kipyin@users.noreply.github.com>"]
license = "MIT"
packages = [
    { include = "impresario", from = "src" },
]

[tool.poetry.dependencies]
python = "^3.7"
pyyaml = "^5.2"
attrs = "^19.3"
pypinyin = "^0.36.0"
tablib = "^0.14.0"

[tool.poetry.dev-dependencies]
black = {version = "^18.3-alpha.0", allows-prereleases = true}
pytest = "^5.3.2"
tox = "^3.14.2"
isort = {version = "^4.3.21", extras = ["pyproject"]}

[tool.poetry.scripts]
songbook = "src.songbook.console:main"

[tool.isort]
atomic = true
multi_line_output = 3

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
